[refactor] ScheduleConverter 위치 변경
변경 사항 :
- ScheduleConverter : 패키지 위치 변경
