+++ src/main/java/com/final_10aeat/domain/manageArticle/service/ManageScheduleService.java
@@ -2,7 +2,7 @@ package com.final_10aeat.domain.manageArticle.service;
 
 import com.final_10aeat.common.exception.UnauthorizedAccessException;
 import com.final_10aeat.domain.manageArticle.dto.request.ScheduleRequestDto;
-import com.final_10aeat.domain.manageArticle.dto.request.util.ScheduleConverter;
+import com.final_10aeat.domain.manageArticle.dto.util.ScheduleConverter;
 import com.final_10aeat.domain.manageArticle.entity.ManageArticle;
 import com.final_10aeat.domain.manageArticle.entity.ManageSchedule;
 import com.final_10aeat.domain.manageArticle.exception.ScheduleMustHaveOneException;
